requirements.txt
appdirs==1.4.3
asgiref==3.2.3
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==2.0.2
django-3-jet==1.0.7
django-ckeditor==5.9.0
django-import-export==2.1.0
django-js-asset==1.2.2
django-material-admin==1.7.0
et-xmlfile==1.0.1
feedparser==5.2.1
filelock==3.0.12
gunicorn==20.1.0
importlib-metadata==1.6.0
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.2
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
psycopg2==2.9.5
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.0
tablib==1.1.0
virtualenv==20.0.18
whitenoise==6.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
